[MOVEONLY] Move unused Merkle branch code to tests
Zcash: Also includes some minor changes from scripted diffs that
occurred between backported PRs.

(cherry picked from commit bitcoin/bitcoin@4defdfab94504018f822dc34a313ad26cedc8255)
